python-modules/requirements.txt
google-generativeai
langchain
langchain-core
langchain-google-genai
pydantic
sqlalchemy
psycopg2-binary
python-dotenv
fastapi
uvicorn
httpx
aiofiles
